Exercice 1

+++ server/Controllers/MarketController.js
@@ -15,7 +15,19 @@ class MarketController{
     getMarket(req, res){
         if(req.params.id){
             const id = req.params.id
+            if(!this.service.isIdExist(id)){
+                throw new Error("This market doesn't exist")
+            }
             const market = this.service.getMarketById(id)
+            res.send(market)
+        } else if(req.params.name){
+            let name = req.params.name
+            const nameArray = name.split("_")
+            name = nameArray.join(" ")
+            if(!this.service.isNameExist(name)){
+                throw new Error("This market doesn't exist")
+            }
+            const market = this.service.getMarketByName(name)
             res.send(market)
         }
     }
@@ -28,9 +40,23 @@ class MarketController{
     deleteMarket(req, res){
         if(req.params.id){
             const id = req.params.id
+            if(!this.service.isIdExist(id)){
+                throw new Error("This market doesn't exist")
+            }
+
             this.service.deleteMarketById(id)
             res.send(`Market ${id} deleted`)
         }
+    }
+
+    updateMarket(req, res){
+        const id = req.params.id
+        if(!this.service.isIdExist(id)){
+            throw new Error("This market doesn't exist")
+        }
+
+        this.service.updateMarket(req, this.service.getMarketById(id))
+        res.send(`Market ${id} updated`)
     }
 }
